Add a pad? column option
It provides a way to turn off all padding for a column, which is useful to prevent
trailing whitespace in the right-most column.  I want this for the --help output produced
by command-line-parser. (Admittedly this is mostly useful for OCD people like me who
highlight trailing whitespace in emacs, but it can also affect copy/paste.)